+++ src/main/java/com/imsweb/datagenerator/hl7/NaaccrHl7DataGenerator.java
@@ -21,6 +21,7 @@ import com.imsweb.datagenerator.hl7.rule.ObservationRequestSegmentRule;
 import com.imsweb.datagenerator.hl7.rule.ObservationSegmentRule;
 import com.imsweb.datagenerator.hl7.rule.PatientIdentifierSegmentRule;
 import com.imsweb.datagenerator.hl7.rule.PatientVisitSegmentRule;
+import com.imsweb.datagenerator.hl7.rule.SpecimenSegmentRule;
 import com.imsweb.layout.LayoutFactory;
 import com.imsweb.layout.hl7.NaaccrHl7Layout;
 import com.imsweb.layout.hl7.entity.Hl7Message;
@@ -59,6 +60,7 @@ public class NaaccrHl7DataGenerator implements DataGenerator {
         _rules.add(new NextOfKinSegmentRule());
         _rules.add(new PatientVisitSegmentRule());
         _rules.add(new CommonOrderSegmentRule());
+        _rules.add(new SpecimenSegmentRule());
         _rules.add(new ObservationRequestSegmentRule());
         _rules.add(new ObservationSegmentRule());
     }